Revert "RE:[STACK-2285][STACK-2263][STACK-2286][STACK-2300][STACK-2305] Added test harness"
